Resolve conflicts
